abstract getMostRelevantCompany
