+++ packages/core/src/modules/job-offers/job-offers.ts
@@ -1,16 +1,15 @@
 import dedent from 'dedent';
-import { type Transaction } from 'kysely';
 import { match } from 'ts-pattern';
 import { z } from 'zod';
 
-import { type DB, db } from '@oyster/db';
+import { db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
 import { job } from '@/api';
 import { JobOfferBullJob } from '@/infrastructure/bull/bull.types';
 import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { getChatCompletion } from '@/modules/ai/ai';
-import { searchCrunchbaseOrganizations } from '@/modules/employment/queries/search-crunchbase-organizations';
+import { getMostRelevantCompany } from '@/modules/employment/companies';
 import { saveCompanyIfNecessary } from '@/modules/employment/use-cases/save-company-if-necessary';
 import { ENV } from '@/shared/env';
 import { fail, type Result, success } from '@/shared/utils/core.utils';
@@ -459,62 +458,7 @@ export async function editFullTimeJobOffer(
   return success(result);
 }
 
-// Helper functions
-
-// TODO: Refactor this to be a shared utility function. It is also used in
-// opportunity.ts.
-/**
- * Finds the most relevant company ID based on the given name.
- *
- * If the company is already in our database, then this function will return the
- * ID of the existing company.
- *
- * Otherwise, this function will query the Crunchbase API, choose the most
- * relevant company, and save it in our database (if it's not already there).
- * Then returns the ID of the newly created company.
- *
- * @param trx - Database transaction to use for the operation.
- * @param companyName - Name of the company to find or create.
- * @returns ID of the company found or created.
- */
-async function getMostRelevantCompany(
-  trx: Transaction<DB>,
-  companyName: string
-) {
-  const companyFromDatabase = await trx
-    .selectFrom('companies')
-    .select('id')
-    .where('name', 'ilike', companyName)
-    .executeTakeFirst();
-
-  if (companyFromDatabase) {
-    return companyFromDatabase.id;
-  }
-
-  const [company] = await searchCrunchbaseOrganizations(companyName);
-
-  if (company && areNamesSimilar(companyName, company.name)) {
-    return saveCompanyIfNecessary(trx, company.crunchbaseId);
-  }
-
-  return null;
-}
-
-/**
- * Checks if two company names are similar by checking if one string is a
- * substring of the other. This does a naive comparison by removing all
- * non-alphanumeric characters and converting to lowercase.
- *
- * @param name1 - First company name.
- * @param name2 - Second company name.
- * @returns Whether the two company names are similar.
- */
-function areNamesSimilar(name1: string, name2: string) {
-  const normalized1 = name1.toLowerCase().replace(/[^a-z0-9]/g, '');
-  const normalized2 = name2.toLowerCase().replace(/[^a-z0-9]/g, '');
-
-  return normalized1.includes(normalized2) || normalized2.includes(normalized1);
-}
+// Helpers
 
 // "Has Edit Permission"
